near_swarm/core/swarm_agent.py
"""
NEAR Swarm Agent Implementation
Extends base NEARAgent with swarm intelligence capabilities
"""

from typing import Dict, List, Optional
from dataclasses import dataclass
import asyncio

from .agent import NEARAgent, AgentConfig
from .consensus import Vote, ConsensusManager


@dataclass
class SwarmConfig:
    """Configuration for swarm behavior"""
    role: str  # Specialized role in the swarm
    min_confidence: float = 0.7
    min_votes: int = 3
    timeout: float = 5.0


class SwarmAgent(NEARAgent):
    """NEAR Protocol Swarm Agent implementation"""
    
    def __init__(self, config: AgentConfig, swarm_config: SwarmConfig):
        """Initialize the swarm agent"""
        super().__init__(config)
        self.swarm_config = swarm_config
        self.consensus_manager = ConsensusManager(
            min_confidence=swarm_config.min_confidence,
            min_votes=swarm_config.min_votes,
            timeout=swarm_config.timeout
        )
        self.swarm_peers: List[SwarmAgent] = []
    
    async def join_swarm(self, peers: List['SwarmAgent']):
        """Join a swarm of peer agents"""
        self.swarm_peers = peers
        await self._announce_presence()
    
    async def _announce_presence(self):
        """Announce presence to swarm peers"""
        for peer in self.swarm_peers:
            try:
                await peer.handle_peer_announcement(self)
            except Exception as e:
                self.logger.error(f"Failed to announce to peer: {e}")
    
    async def handle_peer_announcement(self, peer: 'SwarmAgent'):
        """Handle announcement from a new peer"""
        if peer not in self.swarm_peers:
            self.swarm_peers.append(peer)
    
    async def propose_action(self, action_type: str, params: Dict) -> Dict:
        """Propose an action to the swarm"""
        proposal = {
            "type": action_type,
            "params": params,
            "proposer": self.config.account_id
        }
        
        # Collect votes from peers
        votes = await self.consensus_manager.collect_votes(
            proposal_id=str(hash(str(proposal))),
            agents=self.swarm_peers,
            proposal=proposal
        )
        
        # Reach consensus
        result = self.consensus_manager.reach_consensus(votes)
        return {
            **result,
            "proposal": proposal
        }
    
    async def evaluate_proposal(self, proposal: Dict) -> Dict:
        """Evaluate a proposal based on agent's role and expertise"""
        try:
            # Implement role-specific evaluation logic
            if self.swarm_config.role == "market_analyzer":
                return await self._evaluate_market_conditions(proposal)
            elif self.swarm_config.role == "risk_manager":
                return await self._evaluate_risk(proposal)
            elif self.swarm_config.role == "strategy_optimizer":
                return await self._evaluate_strategy(proposal)
            else:
                return await self._evaluate_general(proposal)
        except Exception as e:
            self.logger.error(f"Proposal evaluation failed: {e}")
            return {
                "decision": False,
                "confidence": 0.0,
                "reasoning": f"Evaluation failed: {str(e)}"
            }
    
    async def _evaluate_market_conditions(self, proposal: Dict) -> Dict:
        """Evaluate market conditions for a proposal"""
        # Example market analysis logic
        try:
            if proposal["type"] == "trade":
                price = await self.check_price(proposal["params"]["token"])
                volume = await self.check_volume(proposal["params"]["token"])
                
                confidence = min(volume / 10000, 1.0)  # Example confidence calc
                return {
                    "decision": volume > 5000 and price > 0,
                    "confidence": confidence,
                    "reasoning": f"Market conditions: Price=${price}, Volume=${volume}"
                }
        except Exception as e:
            return {
                "decision": False,
                "confidence": 0.0,
                "reasoning": f"Market analysis failed: {str(e)}"
            }
    
    async def _evaluate_risk(self, proposal: Dict) -> Dict:
        """Evaluate risk factors for a proposal"""
        # Implement risk evaluation logic
        return {
            "decision": True,
            "confidence": 0.8,
            "reasoning": "Basic risk check passed"
        }
    
    async def _evaluate_strategy(self, proposal: Dict) -> Dict:
        """Evaluate strategy optimization for a proposal"""
        # Implement strategy evaluation logic
        return {
            "decision": True,
            "confidence": 0.9,
            "reasoning": "Strategy parameters within bounds"
        }
    
    async def _evaluate_general(self, proposal: Dict) -> Dict:
        """General proposal evaluation"""
        return {
            "decision": True,
            "confidence": 0.7,
            "reasoning": "Basic validation passed"
        }
    
    async def check_price(self, token: str) -> float:
        """Check token price"""
        # Implement price checking logic
        return 1.0
    
    async def check_volume(self, token: str) -> float:
        """Check token volume"""
        # Implement volume checking logic
        return 10000.0 